requirements.txt
braintree==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
comtypes==1.1.7
coverage==5.0.4
cryptography==2.9
django>=2.2.13
django-jet==1.0.8
django-prettyjson==0.4.1
gunicorn==19.9.0
idna==2.9
paypalrestsdk==1.13.1
Pillow==7.1.2
psycopg2==2.8.3
pycparser==2.20
pyOpenSSL==19.1.0
pytz==2019.1
requests==2.23.0
six==1.12.0
sqlparse==0.3.0
standardjson==0.3.1
urllib3==1.25.8
whitenoise==4.1.3
